rewrite iterator, yet again
Split iterating over an indexedDB into a separate module so that
it can be used as a regular node readable stream. Furthermore
replace the "reopenOnTimeout" option with "snapshot", which defaults
to true.

This makes the iterator iterate over a snapshot of the database, just
like leveldb. This is at the cost of memory though. If back pressure
is needed in order to prevent memory spikes, use snapshot = false.